+++ scripts/update-version.ts
@@ -1,7 +1,10 @@
+import { readFile, writeFile } from 'node:fs/promises'
 import { join } from 'node:path'
 
 // Writes the current package.json version to `./src/version.ts`.
-const versionFilePath = join(import.meta.dir, '../src/version.ts')
-const packageJsonPath = join(import.meta.dir, '../src/package.json')
-const packageVersion = (await Bun.file(packageJsonPath).json()).version
-Bun.write(versionFilePath, `export const version = '${packageVersion}'\n`)
+const versionFilePath = join(import.meta.dirname, '../src/version.ts')
+const packageJsonPath = join(import.meta.dirname, '../src/package.json')
+const packageVersion = JSON.parse(
+  await readFile(packageJsonPath, 'utf8'),
+).version
+await writeFile(versionFilePath, `export const version = '${packageVersion}'\n`)