+++ src/evm/chain/chains.ts
@@ -499,6 +499,11 @@ const xLayer = /* @__PURE__ */ defineEvmChain(xLayerViem, {
   parentChainId: ethereum.chainId,
 })
 
+const robinhood = /* @__PURE__ */ defineEvmChain(robinhoodViem, {
+  key: 'robinhood',
+  shortName: 'robinhood',
+})
+
 // Testnets
 const sepolia = /* @__PURE__ */ defineEvmChain(sepoliaViem, {
   key: 'sepolia',
@@ -631,7 +636,7 @@ export const evmChains = [
   monad,
   megaeth,
   xLayer,
-
+  robinhood,
   sepolia,
   arbitrumSepolia,
   tatara,