+++ src/evm/config/features/operations-multisig.ts
@@ -45,6 +45,7 @@ export const MULTISIG_CHAIN_IDS = [
   EvmChainId.MONAD,
   EvmChainId.MEGAETH,
   EvmChainId.XLAYER,
+  EvmChainId.ROBINHOOD,
 ] as const
 
 export type MultisigChainId = (typeof MULTISIG_CHAIN_IDS)[number]
@@ -94,6 +95,7 @@ export const MULTISIG_ADDRESS: Record<MultisigChainId, `0x${string}`> = {
   [EvmChainId.MONAD]: '0xc4c2fc252f51b3fad6120963acf0d560526abb00',
   [EvmChainId.MEGAETH]: '0xa5b3262d0c239a82d99ecd785e07459f258160b2',
   [EvmChainId.XLAYER]: '0x67056d0429932053d8c78a248d9965f36961fe04',
+  [EvmChainId.ROBINHOOD]: '0x6fa4080b70eed82cf48570dde705026905a64657',
 } as const
 
 export const isMultisigChainId = (