+++ src/evm/config/features/uniswap-v3.ts
@@ -57,6 +57,7 @@ export const UNISWAP_V3_SUPPORTED_CHAIN_IDS = [
   EvmChainId.PLASMA,
   EvmChainId.MONAD,
   EvmChainId.XLAYER,
+  EvmChainId.ROBINHOOD,
   EvmChainId.MEGAETH,
   EvmChainId.ROBINHOOD,
 ] as const
@@ -102,6 +103,7 @@ export const UNISWAP_V3_INIT_CODE_HASH: Record<
   [EvmChainId.PLASMA]: POOL_INIT_CODE_HASH,
   [EvmChainId.MONAD]: POOL_INIT_CODE_HASH,
   [EvmChainId.XLAYER]: POOL_INIT_CODE_HASH,
+  [EvmChainId.ROBINHOOD]: POOL_INIT_CODE_HASH,
   [EvmChainId.MEGAETH]: POOL_INIT_CODE_HASH,
   [EvmChainId.ROBINHOOD]: POOL_INIT_CODE_HASH,
 } as const
@@ -137,6 +139,7 @@ export const UNISWAP_V3_FACTORY_ADDRESS: Record<
   [EvmChainId.PLASMA]: '0xcb2436774c3e191c85056d248ef4260ce5f27a9d',
   [EvmChainId.MONAD]: '0x204faca1764b154221e35c0d20abb3c525710498',
   [EvmChainId.XLAYER]: '0x4b2ab38dbf28d31d467aa8993f6c2585981d6804',
+  [EvmChainId.ROBINHOOD]: '0x1f7d7550b1b028f7571e69a784071f0205fd2efa',
   [EvmChainId.MEGAETH]: '0x3a5f0cd7d62452b7f899b2a5758bfa57be0de478',
   [EvmChainId.ROBINHOOD]: '0x1f7d7550b1b028f7571e69a784071f0205fd2efa',
 } as const