+++ src/evm/currency/deserialize-currency.ts
@@ -1,3 +1,4 @@
+import { createCurrencyDeserializer } from '../../generic/currency/deserialize-currency.js'
 import { EvmNative, type SerializedEvmNative } from './native.js'
 import { EvmToken, type SerializedEvmToken } from './token.js'
 
@@ -7,11 +8,12 @@ function deserializeEvmCurrency(data: SerializedEvmNative): EvmNative
 function deserializeEvmCurrency(
   data: SerializedEvmNative | SerializedEvmToken,
 ) {
-  if (data.type === 'native') {
-    return new EvmNative(data)
-  } else {
-    return new EvmToken(data)
-  }
+  return deserialize({ ...data, type: data.type })
 }
+
+const deserialize = createCurrencyDeserializer({
+  native: EvmNative,
+  token: EvmToken,
+})
 
 export { deserializeEvmCurrency }