+++ src/generic/validate/index.ts
@@ -1,3 +1,4 @@
+export * from './create-address-schema.js'
 export * from './isPromiseFulfilled.js'
 export * from './isPromiseRejected.js'
 export * from './zod.js'