+++ src/stellar/currency/deserialize-currency.ts
@@ -1,7 +1,10 @@
+import { createCurrencyDeserializer } from '../../generic/currency/deserialize-currency.js'
 import { type SerializedStellarToken, StellarToken } from './token.js'
 
 function deserializeStellarCurrency(data: SerializedStellarToken) {
-  return new StellarToken(data)
+  return deserialize(data)
 }
+
+const deserialize = createCurrencyDeserializer({ token: StellarToken })
 
 export { deserializeStellarCurrency }